# apidocs/swagger.yaml
basePath: /
definitions:
  github_com_stockflow_server_internal_response.Envelope:
    properties:
      code:
        description: 成功 0（数字）；失败为模块命名空间错误码字符串（COMMON_* 等）
      data: {}
      details:
        description: 校验失败必带（api.md §4）
      message:
        type: string
      request_id:
        type: string
    type: object
...
info:
  contact: {}
  description: StockFlow（库流智能仓储管理系统）后端 API。统一信封 {code,message,data,request_id,details}（api.md
    §2）。
  title: StockFlow API
  version: "1.0"
paths:
...
  /api/inbounds:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/inbounds
      tags:
      - 采购入库
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/inbounds
      tags:
      - 采购入库
  /api/inbounds/{id}:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/inbounds/:id
      tags:
      - 采购入库
    put:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/inbounds/:id
      tags:
      - 采购入库
  /api/inbounds/{id}/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/inbounds/:id/cancel
      tags:
      - 采购入库
  /api/inbounds/{id}/close:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/inbounds/:id/close
      tags:
      - 采购入库
...
  /api/purchases:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/purchases
      tags:
      - 采购入库
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchases
      tags:
      - 采购入库
  /api/purchases/{id}:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/purchases/:id
      tags:
      - 采购入库
    put:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/purchases/:id
      tags:
      - 采购入库
  /api/purchases/{id}/approve:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchases/:id/approve
      tags:
      - 采购入库
  /api/purchases/{id}/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchases/:id/cancel
      tags:
      - 采购入库
  /api/purchases/{id}/close:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchases/:id/close
      tags:
      - 采购入库
  /api/purchases/{id}/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchases/:id/submit
      tags:
      - 采购入库
  /api/putaway:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/putaway
      tags:
      - 采购入库
  /api/putaway/{id}:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/putaway/:id
      tags:
      - 采购入库
  /api/putaway/{id}/claim:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/putaway/:id/claim
      tags:
      - 采购入库
  /api/putaway/{id}/execute:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/putaway/:id/execute
      tags:
      - 采购入库
  /api/putaway/recommend:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: 推荐库位（§5.3 基础规则经 BinRecommender）
      tags:
      - 采购入库
  /api/quality:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/quality
      tags:
      - 采购入库
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/quality
      tags:
      - 采购入库
  /api/quality/{id}:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/quality/:id
      tags:
      - 采购入库
  /api/quality/{id}/execute:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/quality/:id/execute
      tags:
      - 采购入库
  /api/quality/{id}/start:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/quality/:id/start
      tags:
      - 采购入库
  /api/receipts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/receipts
      tags:
      - 采购入库
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: 收货确认（幂等：Idempotency-Key 头优先于请求体键——plan §7）
      tags:
      - 采购入库
  /api/receipts/{id}:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/receipts/no/:no
      tags:
      - 采购入库
...
securityDefinitions:
  BearerAuth:
    description: 值格式 "Bearer <access_token>"（api.md §6.1）。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"